bind table object to table container and processed data to table node
