+++ src/createTable.js
@@ -14,7 +14,7 @@ export default function createTable(element = 'body', config = {}, controls = nu
 
     thisTable.required_cols = [];
 
-    thisTable.wrap = select(thisTable.div).append('div');
+    thisTable.wrap = select(thisTable.div).append('div').datum(thisTable);
 
     thisTable.events = {
         onInit() {},
